Create ETC_it.tsv
I'm starting the Italian translation. My idea is to translate all the dialogues but keeps the spells, cities' names and monsters name in english. First of all because in Italian it will not sounds good. And then because it will be more easy to communicate and speak in-game with players that have the client in english.

Tell me what do you think about this. Hope to hear you soon. 

I also waiting this game so much and at the moment I never played it myself but keeping my eyes close to all what you doing.